컴포넌트 기반 구조 설계계
