+++ src/controller/render.js
@@ -1,0 +1,9 @@
+/**
+ * HTML Elemnet를 parameter로 page 변경
+ * @param {HTMLElement} page //
+ */
+const render = (page) => {
+  document.body.innerHTML = page;
+};
+
+export default render;
